refactor: 전에 먹던 사료 화면, 사료 검색 화면 feed/view 폴더로 이동
