+++ app/src/main/java/com/project/meongcare/feed/view/OldFeedFragment.kt
@@ -1,17 +1,18 @@
-package com.project.meongcare
+package com.project.meongcare.feed.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.project.meongcare.R
 
-class SearchFeedFragment : Fragment() {
+class OldFeedFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View? {
-        return inflater.inflate(R.layout.fragment_search_feed, container, false)
+        return inflater.inflate(R.layout.fragment_old_feed, container, false)
     }
 }
