+++ app/src/main/res/layout/fragment_old_feed.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".OldFeedFragment">
+    tools:context=".feed.view.OldFeedFragment">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/toolbar_oldfeed"
